Move ExecutionPhase to interpreter module
