+++ generation/src/proof_generation/counting_interpreter.py
@@ -10,8 +10,8 @@ from proof_generation.stateful_interpreter import StatefulInterpreter
 if TYPE_CHECKING:
     from collections.abc import Mapping
 
-    from proof_generation.basic_interpreter import ExecutionPhase
     from proof_generation.claim import Claim
+    from proof_generation.interpreter import ExecutionPhase
     from proof_generation.pattern import ESubst, EVar, MetaVar, Pattern, SSubst, SVar
