+++ generation/src/proof_generation/pretty_printing_interpreter.py
@@ -9,8 +9,8 @@ from proof_generation.proved import Proved
 if TYPE_CHECKING:
     from collections.abc import Callable, Mapping
 
-    from proof_generation.basic_interpreter import ExecutionPhase
     from proof_generation.claim import Claim
+    from proof_generation.interpreter import ExecutionPhase
     from proof_generation.pattern import ESubst, EVar, MetaVar, Pattern, SSubst, SVar
