+++ generation/src/tests/unit/test_memoizing_interpreter.py
@@ -1,4 +1,4 @@
-from proof_generation.basic_interpreter import ExecutionPhase
+from proof_generation.interpreter import ExecutionPhase
 from proof_generation.optimizing_interpreters import MemoizingInterpreter
 from proof_generation.pattern import MetaVar
 from proof_generation.stateful_interpreter import StatefulInterpreter
