+++ generation/src/tests/unit/test_transfer.py
@@ -5,8 +5,8 @@ from typing import TYPE_CHECKING
 
 import pytest
 
-from proof_generation.basic_interpreter import ExecutionPhase
 from proof_generation.claim import Claim
+from proof_generation.interpreter import ExecutionPhase
 from proof_generation.metamath.converter.converter import MetamathConverter
 from proof_generation.metamath.converter.representation import AxiomWithAntecedents
 from proof_generation.metamath.parser import load_database
